+++ Changelog.md
@@ -330,3 +330,6 @@
 	*	Fixed the non-recognition of ExeFS and RomFS files
 	*	Fixed compression recognition
 	*	Added some little scripts I use for the dev, if it can interest someone (util/dev_scripts)
+*	v1.24.59
+	*	Optimized unpack.GFA and unpack.NDS
+	*	Fixed a byte order error in unpack.NARC